chore: move coverage commands into tools
