+++ common/ci/__init__.py
@@ -1,0 +1,1 @@
+"""Shared CI implementation modules."""
