+++ moon.yml
@@ -8,9 +8,11 @@ fileGroups:
     - '.github/**/*'
     - '.moon/**/*'
     - 'apps/**/*'
+    - 'common/**/*'
     - 'docs/**/*'
     - 'infra/**/*'
     - 'scripts/**/*'
+    - 'tools/**/*'
     - 'tests/**/*'
     - '*.md'
     - '*.toml'
