+++ scripts/tests/test_calculate_unified_coverage.py
@@ -1,4 +1,4 @@
-"""Tests for scripts/calculate_unified_coverage.py.
+"""Tests for common.coverage.calculate_unified_coverage.
 Covers unified coverage calculation across backend, frontend, common, and scripts,
 including blacklist pattern exclusions and threshold enforcement.
 """
@@ -9,10 +9,9 @@ from pathlib import Path
 
 import pytest
 
-# Make scripts importable
-sys.path.insert(0, str(Path(__file__).parent.parent))
+sys.path.insert(0, str(Path(__file__).resolve().parents[2]))
 
-import calculate_unified_coverage as cuc  # noqa: E402
+from common.coverage import calculate_unified_coverage as cuc  # noqa: E402
 
 
 # ---------------------------------------------------------------------------
@@ -322,7 +321,7 @@ class TestLcovFileRecords:
 
 
 class TestCalculateUnifiedCoverage:
-    """Unified aggregation combines backend + frontend + common + scripts correctly."""
+    """Unified aggregation combines backend + frontend + common + tools + scripts."""
 
     def _make(self, total, covered):
         return {"total_lines": total, "covered_lines": covered, "coverage_percent": 0}
@@ -369,6 +368,19 @@ class TestCalculateUnifiedCoverage:
         assert result["total_lines"] == 40
         assert result["covered_lines"] == 25
         assert result["breakdown"]["common"] is common
+
+    def test_AC8_13_54_tools_breakdown_is_included_when_present(self):
+        b = self._make(10, 5)
+        fe = self._make(10, 5)
+        s = self._make(10, 5)
+        common = self._make(10, 10)
+        tools = self._make(5, 5)
+
+        result = cuc.calculate_unified_coverage(b, fe, s, common, tools)
+
+        assert result["total_lines"] == 45
+        assert result["covered_lines"] == 30
+        assert result["breakdown"]["tools"] is tools
 
     def test_coverage_close_to_30_pct(self):
         # Simulate production-like scenario: ~30% coverage
