+++ tools/__init__.py
@@ -1,0 +1,1 @@
+"""Command entry points for project tooling."""
