+++ tools/ci/__init__.py
@@ -1,0 +1,1 @@
+"""CI command entry points."""
