+++ tools/ci/check_ci_metrics_contract.py
@@ -1,17 +1,16 @@
 #!/usr/bin/env python3
-"""Compatibility wrapper for common.coverage.merge_lcov."""
+"""Command wrapper for the CI metrics contract."""
 
 from __future__ import annotations
 
 import sys
 from pathlib import Path
 
-ROOT_DIR = Path(__file__).resolve().parents[1]
+ROOT_DIR = Path(__file__).resolve().parents[2]
 if str(ROOT_DIR) not in sys.path:
     sys.path.insert(0, str(ROOT_DIR))
 
-from common.coverage.merge_lcov import *  # noqa: E402,F401,F403
-from common.coverage.merge_lcov import main  # noqa: E402
+from common.ci.metrics_contract import main  # noqa: E402
 
 
 if __name__ == "__main__":  # pragma: no cover
